stock_data_scraper/requirements.txt
# Astro Runtime includes the following pre-installed providers packages: https://www.astronomer.io/docs/astro/runtime-image-architecture#provider-packages
requests
beautifulsoup4
google-cloud-bigquery
pandas
python-dotenv
pyarrow 